Extract shared integration-test helpers to tests/common
The S3 (r2proxy) integration suite carried a verbatim copy of util.py
and its own copies of the launcher plumbing and the case-runner harness,
which the upcoming remote-stripe (toxiproxy) suite would duplicate again.

Move util.py to tests/common/ and add tests/common/harness.py with the
shared pieces: free_port, install_exit_handler (atexit + SIGTERM/SIGINT),
wait_for (readiness poll), and a Suite base class providing ok/notok/uid
and the run() loop. Migrate the S3 suite onto them (Cases subclasses
Suite, make_fixture becomes the setup() hook). Add tests/.gitignore for
__pycache__.
